Create new template

+++ express-template/controllers/contactController.js
@@ -6,7 +6,7 @@ const Contact = require("../models/contactModel");
 //@route GET /api/contacts
 //@access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find({});
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
@@ -23,7 +23,8 @@ const createContact = asyncHandler(async (req, res) => {
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id,
     });
 
     res.status(201).json(contact);
@@ -51,10 +52,10 @@ const updateContact = asyncHandler(async (req, res) => {
         throw new Error("Contact not found");
     }
 
-    // if (contact.user_id.toString() !== req.user.id) {
-    //     res.status(403);
-    //     throw new Error("User don't have permission to update other user contacts");
-    // }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
+    }
 
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
@@ -74,10 +75,10 @@ const deleteContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Contact not found");
     }
-    // if (contact.user_id.toString() !== req.user.id) {
-    //     res.status(403);
-    //     throw new Error("User don't have permission to update other user contacts");
-    // }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
+    }
     await Contact.deleteOne({ _id: req.params.id });
     res.status(200).json(contact);
 });
